Split schedule screen into multiple files
