+++ frontend/src/screens/schedule/types.ts
@@ -1,0 +1,25 @@
+import moment from "moment";
+
+export const ItemTypes = {
+  TALK_30: "talk_30",
+  TALK_45: "talk_45",
+  TALK_60: "talk_60",
+  ALL_TRACKS_EVENT: "all_tracks_event",
+};
+
+export type Slot = {
+  duration: number;
+  hour: moment.Moment;
+  size: number;
+  offset: number;
+};
+
+export type ScheduleItem = {
+  title: string;
+  trackSpan?: number;
+  allTracks?: boolean;
+};
+
+export type Room = {
+  name: string;
+};
